Python Server/requirements.txt
fastapi
uvicorn[standard]
redis>=4.5         
python-dotenv
